requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
atomicwrites==1.4.1
attrs==26.1.0
colorama==0.4.6
flake8==7.3.0
Flake8-pyproject==1.2.4
iniconfig==2.3.0
mccabe==0.7.0
packaging==26.0
pluggy==1.6.0
py==1.11.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.20.0
pytest==7.0.0
tomli==2.4.1
wemake-python-styleguide==1.6.1
